Added DeleteGuidDirectories in DirectoryHelper that deletes any folders that are named after a Guid

+++ src/Horn.Spec.Framework/helpers/DirectoryHelper.cs
@@ -1,10 +1,15 @@
 using System;
 using System.IO;
+using System.Text;
+using System.Text.RegularExpressions;
 
 namespace Horn.Framework.helpers
 {
     public static class DirectoryHelper
     {
+        public const string GuidExpression =
+    @"^(\{{0,1}([0-9a-fA-F]){8}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){4}-([0-9a-fA-F]){12}\}{0,1})$";
+
         public static string GetTempDirectoryName()
         {
             var temp = new DirectoryInfo(Path.Combine(Environment.GetEnvironmentVariable("temp"), Guid.NewGuid().ToString()));
@@ -23,9 +28,21 @@ namespace Horn.Framework.helpers
             if (!tempTreeRootFolder.Exists)
                 tempTreeRootFolder.Create();
 
+            DeleteGuidDirectories(tempTreeRootFolder.Parent);
 
-            foreach (var directory in tempTreeRootFolder.GetDirectories())
+            var revisionDataFile = new FileInfo(Path.Combine(tempTreeRootFolder.FullName, "revision.horn"));
+
+            if (revisionDataFile.Exists)
+                revisionDataFile.Delete();
+        }
+
+        public static void DeleteGuidDirectories(DirectoryInfo root)
+        {
+            foreach (var directory in root.GetDirectories())
             {
+                if (!Regex.IsMatch(directory.Name, GuidExpression))
+                    continue;
+
                 try
                 {
                     directory.Delete(true);
@@ -34,12 +51,7 @@ namespace Horn.Framework.helpers
                 {
                     continue;
                 }
-            }
-
-            var revisionDataFile = new FileInfo(Path.Combine(tempTreeRootFolder.FullName, "revision.horn"));
-
-            if (revisionDataFile.Exists)
-                revisionDataFile.Delete();
+            }            
         }
 
         public static string GetBaseDirectory()
